# [SPARK-40764][SQL] Extract partitioning from children's output expressions