--- a/sql/core/src/main/scala/org/apache/spark/sql/execution/AliasAwareOutputExpression.scala
+++ b/sql/core/src/main/scala/org/apache/spark/sql/execution/AliasAwareOutputExpression.scala
@@ -16,8 +16,10 @@
  */
 package org.apache.spark.sql.execution
 
-import org.apache.spark.sql.catalyst.expressions.{Alias, Expression, NamedExpression, SortOrder}
+import org.apache.spark.sql.catalyst.expressions.{Alias, Attribute, Expression, NamedExpression, SortOrder}
 import org.apache.spark.sql.catalyst.plans.physical.{HashPartitioning, Partitioning, PartitioningCollection, UnknownPartitioning}
+import org.apache.spark.sql.execution.exchange.ReusedExchangeExec
+import org.apache.spark.sql.execution.reuse.ReuseExchangeAndSubquery
 
 /**
  * A trait that provides functionality to handle aliases in the `outputExpressions`.
@@ -29,7 +31,33 @@ trait AliasAwareOutputExpression extends UnaryExecNode {
     case a @ Alias(child, _) => child.canonicalized -> a.toAttribute
   }.toMap
 
+  private lazy val (attributeAliases, nonAttributeAliases) = child.collect {
+    case a: AliasAwareOutputExpression => a.outputExpressions.collect { case a: Alias => a }
+    case r: ReusedExchangeExec => r.getTagValue(ReuseExchangeAndSubquery.ORIGIN_EXCHANGE).map {
+      _.collect {
+        case a: AliasAwareOutputExpression => a.outputExpressions.collect { case a: Alias => a }
+      }.flatten
+    }.getOrElse(Nil)
+  }.flatten.partition(_.child.isInstanceOf[Attribute])
+
+  private lazy val nonAttributeAliasesMap =
+    nonAttributeAliases.map(a => a.toAttribute.canonicalized -> a.child).toMap
+
+  private lazy val attributeAliasesMap =
+    attributeAliases.map(a => a.child.canonicalized -> a.toAttribute).toMap
+
   protected def hasAlias: Boolean = aliasMap.nonEmpty
+
+  protected def extractFromChildren(exp: Expression): Expression = {
+    exp.transformDown {
+      case a: Attribute => nonAttributeAliasesMap.getOrElse(a.canonicalized, a)
+    }.transformDown {
+      case a: Attribute => a
+      case e: Expression => e.transformDown {
+        case a: Attribute => attributeAliasesMap.getOrElse(a.canonicalized, a)
+      }
+    }
+  }
 
   protected def normalizeExpression(exp: Expression): Expression = {
     exp.transformDown {
@@ -47,7 +75,10 @@ trait AliasAwareOutputPartitioning extends AliasAwareOutputExpression {
     val normalizedOutputPartitioning = if (hasAlias) {
       child.outputPartitioning match {
         case e: Expression =>
-          normalizeExpression(e).asInstanceOf[Partitioning]
+          val newOutputPartitioning = normalizeExpression(e).asInstanceOf[Partitioning]
+          val extractedOutputPartitioning =
+            normalizeExpression(extractFromChildren(e)).asInstanceOf[Partitioning]
+          PartitioningCollection(Seq(newOutputPartitioning, extractedOutputPartitioning))
         case other => other
       }
     } else {
@@ -57,7 +88,7 @@ trait AliasAwareOutputPartitioning extends AliasAwareOutputExpression {
     flattenPartitioning(normalizedOutputPartitioning).filter {
       case hashPartitioning: HashPartitioning => hashPartitioning.references.subsetOf(outputSet)
       case _ => true
-    } match {
+    }.distinct match {
       case Seq() => UnknownPartitioning(child.outputPartitioning.numPartitions)
       case Seq(singlePartitioning) => singlePartitioning
       case seqWithMultiplePartitionings => PartitioningCollection(seqWithMultiplePartitionings)
@@ -83,7 +114,13 @@ trait AliasAwareOutputOrdering extends AliasAwareOutputExpression {
 
   final override def outputOrdering: Seq[SortOrder] = {
     if (hasAlias) {
-      orderingExpressions.map(normalizeExpression(_).asInstanceOf[SortOrder])
+      orderingExpressions.map { sortOrder =>
+        val newSortOrder = normalizeExpression(sortOrder).asInstanceOf[SortOrder]
+        val extractedSortOrder =
+          normalizeExpression(extractFromChildren(sortOrder)).asInstanceOf[SortOrder]
+        newSortOrder.copy(sameOrderExpressions = newSortOrder.sameOrderExpressions :+
+          extractedSortOrder.child)
+      }
     } else {
       orderingExpressions
     }